+++ src/metadata-args/EmbeddedMetadataArgs.ts
@@ -19,6 +19,12 @@ export interface EmbeddedMetadataArgs {
     readonly isArray: boolean;
 
     /**
+     * Prefix of the embedded, used instead of propertyName.
+     * If set to empty string, then prefix is not set at all.
+     */
+    readonly prefix?: string;
+
+    /**
      * Type of the class to be embedded.
      */
     readonly type: ((type?: any) => Function);