+++ src/naming-strategy/NamingStrategyInterface.ts
@@ -22,7 +22,7 @@ export interface NamingStrategyInterface {
     /**
      * Gets the embedded's column name from the given property name.
      */
-    embeddedColumnName(embeddedPropertyName: string, columnPropertyName: string, columnCustomName?: string): string;
+    embeddedColumnName(prefixes: string[], columnPropertyName: string, columnCustomName?: string): string;
 
     /**
      * Gets the table's relation name from the given property name.